remove error if supoort dir does not exist

+++ bin/build.ps1
@@ -34,9 +34,12 @@ if (-Not $skiptemplates) {
 <# Cleanup of old files #> 
 Write-Host "Cleaning before new build" -ForegroundColor Green
 & stack clean
-Remove-Item "$deckerdir\resource\decker\support\vendor" -Recurse -Force -ErrorAction Continue
-Remove-Item "$deckerdir\public" -Recurse -Force -ErrorAction Ignore 
-
+if (test-path "$deckerdir\resource\decker\support\vendor") {
+    Remove-Item "$deckerdir\resource\decker\support\vendor" -Recurse -Force -ErrorAction Continue
+}
+if (test-path "$deckerdir\public"){
+    Remove-Item "$deckerdir\public" -Recurse -Force -ErrorAction Ignore 
+}
 
 
 Write-Host "Starting build of standalone binary" -ForegroundColor Green
